backend/requirements.txt
# Core API
fastapi==0.95.2
uvicorn==0.24.0
pydantic==1.10.13
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Machine Learning - MUST BE INSTALLED for Full Comprehensive Analysis
transformers==4.36.0
torch>=2.2.0
sentencepiece>=0.2.1
protobuf>=4.0.0

# NLP and Linguistic Analysis
nltk==3.8.1
spacy==3.7.2
numpy==1.26.2

# External Verification Logic
duckduckgo-search==4.2
validators==0.22.0
google-generativeai==0.8.6

# Legacy/Supplementary (Keep for model compatibility if needed)
scikit-learn==1.3.2
pandas==2.1.3
joblib==1.3.2
aiohttp==3.9.1